Remove GeneralDiscoverable flag
According to the BLE spec (Volume 3, Part C, Section 9.1.1.2), the GeneralDiscoverable flag shouldn't be set for Broadcast devices (which Eddystone should apparently be, when in beacon-advertising mode).

As far as I can tell, this is just used for filtering scan results on the host. I've tested it on Android and iOS and it still works.